requirements.txt
comtypes==1.4.15
numpy==2.2.6
pillow==12.1.0
pypiwin32==223
pyttsx3==2.99
pywin32==311
